src/tests/adapters/endpoints/test_factory_api_model.py
import factory
from app.frameworks_and_drivers.api_models import URLPayload


class URLPayloadFactory(factory.Factory):
    long_url = factory.Faker("url")

    class Meta:
        model = URLPayload
