pyproject.toml
[tool.ruff]
select = [
    "C4", # flake8-comprehensions
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
    "UP", # pyupgrade
]


unfixable = ["B"]

fix = true
line-length = 100
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402"]


[tool.poetry]
name = "linkminimizer"
version = "0.1.0"
description = ""
authors = ["paveL1boyko <pave1.boyk0ok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
aioredis = "^2.0.1"
invoke = "^2.2.0"
motor = "^3.3.1"
pytest = "^7.4.2"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.1" }
factory-boy = "^3.3.0"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
ruff = "^0.0.292"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
